update propTypes, refoctor
