+++ react/features/riff-platform/components/Dashboard/Metrics/Interruptions.js
@@ -13,18 +13,15 @@
  *            MIT License (see https://opensource.org/licenses/MIT)
  *
  * ******************************************************************************/
-
-import { metricsRedux, GraphDatasetTypes, GraphTypes } from '@rifflearning/riff-metrics';
+import { GraphDatasetTypes, GraphTypes } from '@rifflearning/riff-metrics';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 
 import { EventTypes } from '../config';
+import { metricGraphLoaded, getSelectedMeeting, getMetricDataset, getDatasetStatus } from '../utils';
 
 import { StackedBarGraph } from './StackedBarGraph';
-
-const { metricGraphLoaded } = metricsRedux.actions;
-const { getSelectedMeeting, getMetricDataset, getDatasetStatus } = metricsRedux.selectors;
 
 /* ******************************************************************************
  * Interruptions                                                               */ /**
@@ -48,7 +45,7 @@ class Interruptions extends React.PureComponent {
         /** The request status of the graphDataset */
         meeting: PropTypes.shape({
             _id: PropTypes.string.isRequired,
-            participants: PropTypes.instanceOf(Set).isRequired
+            participants: PropTypes.instanceOf(Map).isRequired
         }),
 
         /** sets a graphical rendering status for a graph type to loaded */
