+++ react/features/riff-platform/components/Dashboard/utils.js
@@ -1,0 +1,35 @@
+import { riffUtils, metricsRedux } from '@rifflearning/riff-metrics';
+
+const { logger } = riffUtils;
+const { RequestStatus } = metricsRedux.constants;
+const {
+    loadMetricDatasets,
+    loadUserMeetings,
+    selectMeeting,
+    metricGraphLoaded
+} = metricsRedux.actions;
+const {
+    getSelectedMeeting,
+    getMetricDataset,
+    getDatasetStatus,
+    getAreAllChartsLoaded,
+    getUserMeetings,
+    getUserMeetingsError,
+    getUserMeetingsStatus
+} = metricsRedux.selectors;
+
+export {
+    logger,
+    RequestStatus,
+    loadMetricDatasets,
+    loadUserMeetings,
+    selectMeeting,
+    metricGraphLoaded,
+    getSelectedMeeting,
+    getMetricDataset,
+    getDatasetStatus,
+    getAreAllChartsLoaded,
+    getUserMeetings,
+    getUserMeetingsError,
+    getUserMeetingsStatus
+};
